+++ data/data.csv
@@ -1,10 +1,10 @@
-1,When do doctors get angry?,When they run out of patients.
-2,What's the object-oriented way to become wealthy?,Inheritance
-3,"Dad, can you put my shoes on?",I don't think they'll fit me.
-4,Why do C# and Java developers keep breaking their keyboards?,Because they use a strongly typed language.
-5,Knock-knock.,A race condition. Who is there?
-6,Why are mummys scared of vacation?,They're afraid to unwind.
-7,Where do hamburgers go to dance?,The meat-ball.
-8,What musical instrument is found in the bathroom?,A tuba toothpaste.
-9,If you boil a clown...,Do you get a laughing stock?
-10,What do you call cheese by itself?,Provolone.
+1,Did you know you should always take an extra pair of pants golfing?,Just in case you get a hole in one.
+2,Why does Waldo only wear stripes?,Because he doesn't want to be spotted.
+3,Did you hear about the Mexican train killer?,He had loco motives
+4,What did the ocean say to the shore?,"Nothing, it just waved."
+5,What’s the difference between an African elephant and an Indian elephant?,About 5000 miles.
+6,Did you hear about the guy who invented Lifesavers?,They say he made a mint.
+7,What is a tornado's favorite game to play?,Twister!
+8,How do you make a tissue dance?,You put a little boogie on it.
+9,What do I look like?,A JOKE MACHINE!?
+10,What kind of bagel can fly?,A plain bagel.
