+++ model/jokes.go
@@ -1,0 +1,7 @@
+package model
+
+type Joke struct {
+	ID        int    `json:id`
+	Setup     string `json:setup`
+	Punchline string `json:punchline`
+}
