+++ router/router.go
@@ -1,11 +1,11 @@
-package server
+package router
 
 import (
 	"log"
 	"net/http"
 
-	"academy/handlers/fetch"
-	"academy/handlers/jokes"
+	"academy/controller/jokes"
+	"academy/controller/update"
 
 	"github.com/gorilla/mux"
 )
@@ -13,12 +13,10 @@ import (
 // InitServer will execute in the defined port
 func InitServer() {
 
-	jokes.Load()
 	router := mux.NewRouter().StrictSlash(true)
-	router.HandleFunc("/", jokes.IndexRoute)
 	router.HandleFunc("/api/v1/jokes", jokes.GetJokes).Methods("GET")
 	router.HandleFunc("/api/v1/jokes/{id}", jokes.GetOneJoke).Methods("GET")
-	router.HandleFunc("/api/v1/fetch", fetch.GetData).Methods("GET")
+	router.HandleFunc("/api/v1/new-jokes", update.GetData).Methods("GET")
 
 	log.Fatal(http.ListenAndServe(":3000", router))
 }
