+++ services/updatefile/updatefile.go
@@ -1,22 +1,13 @@
-package fetch
+package updatefile
 
 import (
+	"academy/model"
 	"encoding/csv"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"net/http"
 	"os"
 	"strconv"
 	"strings"
 )
-
-//Joke Struct
-type Joke struct {
-	ID        int    `json:id`
-	Setup     string `json:setup`
-	Punchline string `json:punchline`
-}
 
 func readCsvFile(filePath string) [][]string {
 	f, err := os.Open(filePath)
@@ -34,32 +25,17 @@ func readCsvFile(filePath string) [][]string {
 	return records
 }
 
-//GetData Updates the CSV from the target API
-func GetData(w http.ResponseWriter, r *http.Request) {
-	url := "https://official-joke-api.appspot.com/random_ten"
-	request, error := http.NewRequest("GET", url, nil)
-	if error != nil {
-		fmt.Println(error)
-	}
-	response, _ := http.DefaultClient.Do(request)
-
-	defer response.Body.Close()
-	data, _ := ioutil.ReadAll(response.Body)
-
-	// Unmarshal JSON data
-	var jsonData []Joke
-	json.Unmarshal([]byte(data), &jsonData)
-
+//UpdateFile the CSV from the target API
+func UpdateFile(jsonData []model.Joke) {
 	pwd, err := os.Getwd()
 	csvFile, err := os.OpenFile(pwd+"/data/data.csv", os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
 		fmt.Println(err)
 	}
-
 	defer csvFile.Close()
 	records := readCsvFile(pwd + "/data/data.csv")
-	counter := len(records) + 1
 
+	counter := len(records) + 1
 	writer := csv.NewWriter(csvFile)
 
 	for i, usance := range jsonData {
@@ -70,8 +46,5 @@ func GetData(w http.ResponseWriter, r *http.Request) {
 		writer.Write(row)
 	}
 
-	// remember to flush!
 	writer.Flush()
-	w.Header().Set("Content-type", "application/json")
-	w.WriteHeader(http.StatusOK)
 }
